Use JetpackStatsWidgetsCore framework in Jetpack widgets extension
Also replace existing test for the `WidgetUrlSource` that added the file
to the WordPress target in order to test it with one that imports the
package in the test target (no overhead in the app) to access the type
that is being used.

This is still boardline inappropriate, given that the unit tests for the
app target should cover the app target behavior, but it's the best we
can do without further extracting concepts such as `DeepLinkSource` in a
dedicated library.
