--- WordPress/JetpackStatsWidgets/Extensions/URL+WidgetSource.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-enum WidgetUrlSource: String {
-    case homeScreenWidget = "widget"
-    case lockScreenWidget = "lockscreen_widget"
-}
-
-extension URL {
-    func appendingSource(_ source: WidgetUrlSource) -> URL {
-        var components = URLComponents(url: self, resolvingAgainstBaseURL: false)
-        var queryItems = components?.queryItems ?? []
-        let sourceQuery = URLQueryItem(name: "source", value: source.rawValue)
-        queryItems.append(sourceQuery)
-        components?.queryItems = queryItems
-        return components?.url ?? self
-    }
-}
