+++ WordPress/JetpackStatsWidgets/Views/StatsWidgetsView.swift
@@ -1,3 +1,4 @@
+import JetpackStatsWidgetsCore
 import SwiftUI
 import WidgetKit
